Move player_has_bones() into actor.
Map the monster version to mons_skeleton(), which is the closest
equivalent on that side. It's currently not used directly anywhere,
but it might be useful at some point for an eventual spell that only
affects living monsters with bones.

Whether monsters have hair and/or feet doesn't seem to affect much on
that side, so those functions are still player-specific. (Unless there
are eventual spells to e.g. strangle monsters with their own hair, or
knock their feet out from under them.)