Don't expire Divine Shield with time (elliptic)
When reworking this ability, I'd kept a small duration on top of the
limited block charges with the aim to keep the scope of the ability to a
single combat. But it's been pointed out that this makes it possible to end
a fight with blocks remaining, walk over to a second, and have the effect
expire right as you're about to get hit.

Since there's a non-trivial piety cost to the ability, we shouldn't need to
worry about pre-buffing, so now your block charges simply stay around
indefinitely until used.

(This converts DUR_DIVINE_SHIELD to now be the number of remaining block
charges rather than aut, and removes the use of ATTR_DIVINE_SHIELD
entirely)