Make the Orb complain if any method moves it (ge0FF)
It was possible to use hellfire mortar's temp lava to slowly shift the Orb
of Zot all the way to the stairs before picking it up (by creating lava
under it, causing it to be displaced by one tile). Other less practical
methods of moving it include Tomb cards and Summon Forest.

Since the Orb complains about apportation, it probably ought to complain
about this as well. Now any terrain change that triggers a shift which
moves the orb will have a similar effect as apportation being cast upon it.