Indicate whether species/forms have eyes.
In species, make a new species flag for it. The player species and forms
that are immune to dazzling/blinding (vine stalkers, trees, and fungi)
are now flagged as having no eyes, and that quality is now what gives
them such immunity.

The code assumes elsewhere that all species have eyes; this will be
adjusted.