Try to clean up dazzling/blindness checks.
Move them to actor, and apply them consistently to both player and
monster as much as possible: dazzling checks everywhere, except for
atropa darts, which explicitly do blinding checks.

On the player side, resistance to being dazzled is the same as
resistance to being blinded, to keep dazzling/blinding as available to
players as possible.

This means that vine stalker players, as well as players in tree or
fungus form, can no longer be dazzled/blinded, but all other players
can.