Merge Lesser/Greater Servant abilities into Infernal Servant
As part of the Mahkleb revamp, both demon summoning abilities are being
merged together. Infernal Servant is still obtained at 3*, is priced
between the two previous abilities, and the quality of demon summoned is
now strongly influenced by invocations (and sorted roughly in order of
strength, so that it requires more invocations to get executioners and
balrugs than other 2s, for instance.)

The variety of demons it can summon is expanded to include red devils, ice
devils, rust devils, soul eaters, sun demons, and sixfirhy, while losing
neqoxecs (which were very weak as allies and unpleasant as enemies).

Hostile demons are no longer produced by failing to use this ability, but
instead there is a flat 1-in-6 chance that any use of it will produce a
durably summoned hostile demon *alongside* your friendly one - along with a
small increase in quality of the friendly demon summoned at those times.
The hope is this preserves some of the 'deal with a devil' flavor of the
ability while making it less correct to 'use at the start of battle and
then ?tele if you get a bad roll'.