Add ARTP_TROG_LOUD, increased melee noise and BiA health
(Doesn't yet generate - will be used in Trog gifts.)

This property, displayed as TrogLoud, makes your melee attacks 2.5 times as
loud (which is very loud indeed for a sufficiently damaging attack, though
not as loud as Gong). It also makes your Brothers in Arms Trog allies have
20% extra health - the flavor here is a little vague (your noisiness channels
the power of Trog to them or something), but the idea is that your melee
noise makes it more likely that monsters will wander over and join the fight
at a bad time, but Brothers in Arms is a good answer to this.