Add code framework for gaining Infernal Marks
This adds the basic backend for gaining Makhleb's new capstone gift - one
of a selection of Makhleb-specific mutations that grant powerful passive
effects and/or change how your Makhleb invocations work.

At 6*, the player will be offered a choice of 3 different Marks (out of
however many exist - currently planned at 9). Selecting one of these will
permanently brand the player with it (injuring them greatly in the
process...) and give them its effect for as long as they worship Makhleb.
The mark will persist as a mutation, even if they abandon the god, but will
not actually function unless they start worshipping them again (this
prevents any possibility of abandoning to reroll your mutation
possibilities, and it's also good flavour that you have to live with the
consequences of your actions here.) Forms which normally suppress mutations
will not suppress these.

(They also cause silver vulnerability - despite otherwise being
perma-mutations, they are clearly demonic in origin and ought to count.)

None of the marks included in this commit *do* anything yet; that will come
in future commits.