Remove an unneeded rampage terrain prompt (CarefulOdds)
Since a0b39eb048, rampage has been issuing a check_moveto terrain prompt
if the player does a rampage double-move across a tile that pings
check_moveto, even if that move would end on a "safe" tile. (E.g. rampaging
through a poison cloud into an open tile.) This seems unnecessary as long
as we're checking the player's final position, which we are.