New Xom action: fake Shatter
An rather rare action meant to provide more uniquely Dungeon-altering
material as Xom alone can abuse, in the veins of c695c3f and f441f4b. This
action has a few changes from Shatter: it has much worse odds to destroy
most features beisdes doors or diggable objects (enough to make it much
more likely Xom will malmutate, banish, or teleport people hoping Xom will
cut through a vault for them), it leaves behind piles of debris in the
rock or stone features it destroys, and most importantly, it doesn't
actually do any damage- but it still makes one of the loudest noises in
the game.

The singular nature of the random destruction is somewhat neutral, really-
thinner corridors become wider, opener areas may produce new corridor
veins out of them, rooms may break open, and it's slow for all but the
last of these to matter much in one combat due to the slowly withering
nature of debris- but the extreme of a proper (fake) shatter still marks
it as a net bad effects. It's a reasonable joke, at any rate, for Xom to
do something unique.