Fix an obscure crash with tgw_kikubaaqudgha
This vault used lua to place a miasma cloud, killing a caged human and
'raising' them as a zombie. A script called upon the human dying deleted
the cloud they were standing in, but due to a change introduced in c0b93f29f480c8b389dd038c0c571e0ef521d79d
it was now possible for them to die before _actor_apply_cloud was finished
looking at that very cloud which it still maintained a reference to,
resulting in undefined behavior.

Despite what the lua comment says, the presence of the cloud does not
actually prevent placing a zombie there (perhaps it did when the comment
was written 13 years ago), so let's just not delete it.