Clear out the Clear Minds Boutique
There were two shops using the gimmick of having randarts with
properties that can't normally generate: Clear Minds Boutique and
One-of-a-Kind Rings. This is a bit much for a gimmick that's already on
the edge of what these people let me get away with. Clear Minds was made
entirely as a proof-of-concept while gammafunk was implementing artprop
specification syntax, while One-of-a-Kind Rings already existed before
then and has a stronger theme. Also, by only generating one such
randart, One-of-a-Kind Rings is a little less egregious with its
gimmickry. Accordingly, Clear Minds is getting cut.

While I'm rummaging around in shops.des, One-of-a-Kind Rings is also
getting an inventory tweak: the HP and MP artprops get their higher
ends toned down somewhat.