Mahkleb capstone: Mark of Annihilation
Replaces Infernal Servant with Globe of Annihilation, which launches an
iood-like projectile that explodes in a large blast of the player's
signature destruction. It moves at speed 15 (so about half the speed of an
iood), and the size of the explosion increases with the distance travelled:
no explosion (and only 1/2 damage) at point-blank range, radius 1 at 1 tile
travelled, 2 at 2-3 tiles travelled, and finally 3 at 4+ tiles travelled.

Part of the logic of this replacing servant is that while this is a
destruction-flavored ability, I didn't want it to completely eat up having
a 'free' conjurations button of flexible elements with something much more
expensive and less flexible. And it wouldn't play very well with servants
anyway, so the loss might not be that meaningful. (I considered it simply
being a 3rd ability, but at the moment am saving that for something a bit
flashier; I could be convinced otherwise, with testing.)