Makhleb capstone: Mark of Haemoclasm
This mark gives a passive 10% chance whenever the player kills a monster
themselves for it to explode in a violent rain of gore, damaging enemies in
a 3x3 area. The damage the explosion does is proportional to the max hp of
the thing killed (and cannot harm the player or their allies). If an enemy
dies to this explosion, they instead have a 100% chance to explode,
allowing for occasional dramatic chain explosions (especially if the player
has been using AoE damage sources up to this point, which likely gives it
additional synergy with axes and conjurations).