Makhleb capstone: Mark of the Celebrant
Whenever a player with this mark is reduced to 50% or less HP, they
unleash a flurry of blood arrows aimed randomly at nearby enemies (at least
one per enemy, though if there are fewer than 8, it may shoot at the same
enemy up to twice, with the remaining shots scattered randomly).

Rather than a time-based cooldown, the passive recharges immediately when
the player returns to full HP again - which in some situations might even
happen due to Makhleb's healing in a protracted battle. To help this
slightly, kills made with these blood arrows will always trigger Makhleb's
healing passive.

Given that it is effectively 'easier' to trigger and recharge with less max
HP, and provides an immediately response to help stabilize after spike
damage, it is expected this will be relatively more appealing to lower HP
/ EV-based characters than Marks which rely on paying flat HP instead.