Mahkleb capstone: Mark of Execution
Allows a fragment of an executioner to rest dormant inside oneself.
Whenever you kill an XP-granting enemy in melee, there is a 1-in-8 chance
for it to awaken and manifest their iconic whirl of blades around the
player. For the duration, the player gains rampaging and a very good aux
attack (whose base damage scales with XL) which also triggers upon the
player dodging (like minotaur retaliatory headbutts).

The duration of this status effect is very brief, but it does not expire
for actions that the player spent attacking (allowing one to keep it up
indefinitely as long as they keep swinging).