Make tension track more statuses from both players and monsters
For the most part, the tension formula hasn't been adjusted since e11a346
fourteen(!) years ago. It's an extremely imprecise formula in a multitude
of ways- it doesn't know the actual attack or spell ranges of any given
monsters, anything about monster support capacities, the path and presence
of spell components like clouds or ramparts or prisms, the players' and
the monsters' resistances, and so on and so forth. It somewhat works as a
very simple way to track the density of hostile monsters present versus
one's own health, but would be quite burdensome to make be actually
reasonably good at guessing much more than that.

As a result, we've tried to use it for less and less actually meaningful
roles over time. Currently it's most prominent for reasonably informing
Xom of when to act, whether or not to forbid Beogh's apostle challenges,
and if a vault warden should close a given door or not, all of which are
reasonably relevant enough that they should reasonably work regardless of
how little we'd want to use tension elsewhere.

While I'm not pushing for making tension actually become more unweighty
and lengthy, it's currently already supporting a number of enchantments
on players and monsters, and an update for all of the particularly
dramatic new effects added over the past 14 years should be reasonable
to quickly work into what already exists. (For reference's sake, slow
counts for x1.5 tension, being netted or webbed is x2 tension, and
of course none of this matters if there's nothing hostile in sight.)

 * Players with corrosion get a scaling multiplier for each point of
   corrosion they have on them.
 * Players at 10% or lower mp adds a x1.125 multiplier.
 * Player mesmerization and fear adds a x1.2 multiplier.
 * Players that are fragile, have the -Potion status, or the -Scroll
   status get a x1.33 multiplier.
 * Players that are marked, attractive, unable to cast due to Sap Magic,
   in bat form, or in wisp form get a x1.5 multiplier.
 * Players in fungus, pig, or tree form get a x1.66 multiplier.
 * Monsters that have brilliance or are armed by an undying armoury get
   a x1.25 multiplier.

(The only particularly controversial point here is most likely MP, but
most gods use MP for invocations even on non-spellcasters, tension is
fuzzy anyway, and most reliable mp drainers should have some reasonable
threat to them also, so it's probably fine. There's a long list of
absent bits here- sticky flame and poison are the most conspicuously
absen-, but this probably shouldn't go for much else not added here,
since most of the above are player-option-limiting in some obvious
dramatic fashion or another.)

Since most of what relies on tension is both invisible and randomized for
most players, this shouldn't be too concerningly gameable or complex.
Zapping one's self with polymorph, drinking attraction, and drinking
lignification are all now technically options for grabbing Xom's
attention or delaying an apostle challenge, but this is quite
unlikely to make things net better than if one did not in fact do any of
these things. The player also could have just hurt themselves instead
to get the same rough circumstances change. What this does mean,
however, is that Xom should feel fairly more proactive in a wider number
of scenarios and Beogh will send challengers in less awful situations,
which is plenty enough gain for both gods to probably be worth a page's
worth of easily-adjustable and easily-removable checks.

(Others may quibble about these exact values. Future commits may change
any of these multipliers' strength.)