Mahkleb capstone: Mark of Atrocity
This Mark increases both the power and cost of the player's Unleash
Destruction ability each time it is used consecutively (ie: without doing
anything else inbetween). Upon using it a 4th time, it will also unleash
4 additional empowered bolts of destruction at random distant targets and
reset this buff.

Some of the additional cost is in flat HP, making this slightly better in
relative terms on high HP species.