Increase Coglin gizmo-related vocabulary
After consulting with a consortium of Coglin engineers, we have
learned a great deal about their technology, and accordingly we
have been able to expand the technical language used in Crawl to
describe their magnificent gizmos.

In other words, I added a few entries to gizmo.txt, including both
real-world technical jargon and technobabble memes. Just a few.

Some of the names thus generated may seem redundnant -- it's possible
to get, say, an "autoautomatic autolator" -- but remember that
Coglin technology is very advanced, and their terminology must be
very precise to include all the nuances of meaning, which sometimes
appears as redundancy to lesser minds. Thank you for understanding.