New Xom action: wave of despair
Xom already copies a variety of potions as one of the most common effects
granted to the player, so why not also some of the more tactically
interesting scrolls? In particular, one of the tension spells is using
a much weaker version of a consumable, so it's a good idea to make it
a little more recognizably distinct.

This new rare beneficial action is meant to be dramatic in a different
sense than prior debuffs or destruction by combining a few effects
together at once. After Xom tells a spooky story, inanimate skeletons pop
out, and the ensuing wave of horror halves will, then drains, then fears
everything else in sight. The Will/2 state lasts fairly longer than the
fear, making it so stabbers can use other hexes if possible once they
start chasing the player again, and also applies to the player, so there's
still some risk around the right monsters as Xom should always readily
wield.

(To help make this a bit more visible, I added a quick Will/2 status icon
derived from the current scroll of vulnerability icon.)

Consequentially, Cause Fear has been replaced in the Xom tension spells
list with Dazzling Flash (which is weak, but not much weaker than Small
Mammals already is) and Summon Blazeheart Golem (which has a cute
"suddenly handed a bomb" mechanical capacity).