Don't be content with shifting items into closed doors
When trying to relocate an item because the terrain beneath it changed, any
'traversable' square that doesn't destroy automatically destroy items (ie:
lava) was considered an adequate place to put them. But closed doors are
considered traversable only because they *can* be opened, not because they
aren't solid right now.