Place the Orb of Zot on an untransformable dais
When b22d4219adc6 was done to address Hellfire Mortar cheese, I didn't
fully consider all the different effects that might cause the orb to be
displaced without the player anticipating this. While it's arguable that
Xom starting the orb run without warning is just Xom being Xom, it also
seemed likely that casting Summon Forest near the orb could randomly do so,
as well as Imprisoning a monster adjacent to the orb, and likely other
things.

Starting the orb run unexpectedly is a nasty surprise, so instead of doing
that whenever the orb is displaced by (possibly unintentional) terrain
changes, let's just veto all terrain changes beneath the orb's original
location entirely (by giving it a special pedestal feature that is immune
to such things).