Slightly improve ghost visibility
Player ghosts were kind of difficult to see in tiles and webtiles,
especially Op ghosts, which tended to easily blend with their vault
backgrounds. This commit slightly increases the alpha of ghost tiles:
r-i had mentioned possibly adding an aura to these to improve visibility
as well. More to come in a future commit?