Add @random_body_part_external@.
This is the same as @random_body_part@, except that it leaves out blood
and bones. (If a species with external blood and/or bones is added, it
will obviously need to be adjusted for them.)

Make Xom use it in clone drop messages, so that mirrors don't get
smashed over parts of the player that should not be on the outside.