fix: Draw ghost backgrounds in webtiles
This was left out of 767fb09e7d. Thanks again to index for her tiles work
here, this seems like a nice improvement!

We're missing an unusual background tile for ghosts, but there should
be no way for ghosts to have unusual items with the current code anyway:
I've defaulted unusual ghosts to have the regular unusual highlight just
to let this codepath work properly in the future.