Makhleb capstone: Mark of the Fanatic
The most flashy and complex mark by far, this grants the player the
Vessel of Slaughter ability - a high piety-cost ability that transforms
them into a demonic avatar of Makhleb and unleashes a large blast of
damnation around themselves as they do so.

While in this form, the player has demonic resistances (including torment
immunity), will invulnerability, takes 1/3rd less damage from all sources,
recieves doubled healing from kills, and has a bonus to all outgoing damage
proportional to how low health they were when they activated the ability
(with the expected healing from damnation kills encouraging the player to
'push their luck' a little).

But in exchange, once the transformation wears off, the player will be
dragged into the Crucible of Flesh and made to suffer and slay in payment.

The Crucible is a medium-size portal arena that continually spawns waves of
hostile demons (scaled to the player's level and also how many runes they
have collected beyond 3) and paralyzed mortal victims. The player must earn
their exit by finding and slaughtering enough of these victims (killing
demons also counts for half as much, but the faster the player can manage
it by taking out 'helpless' targets, the less bad things will be for them.)
Killing a mortal victim will immediately spawn another somewhere else on
the map, while demons instead arrive at semi-regular intervals (up to a
cap).

Kills in the Crucible grant no xp or items, but still count for Makhleb's
heal-on-kills passive, to help allow the player to sustain themselves
(since resting will often be difficult here). There are also a small number
of single-use teleport traps scattered around which the player can use for
short-term escapes.

In some ways, the gameplay of the Crucible is reminiscent of being
banished, but should still be sufficient different as to be distinct - the
map is pre-revealed and stable, it has walls instead of being endless, the
monster set is more controlled, ?teleport works the same as usual, etc.

A fundamental flavor aspect of Mahkleb's Marks is about accepting hellish
bargains at the cost of oneself, and I felt like at least one of them
should lean as hard as possible into classical 'deal with the devil'
flavour: tremendous power at a cost one may come to regret later. And
Vessel of Slaughter's power is undeniable, and its cost unique. I hope
people find it fun! (I am sure various numerical knobs will need tweaking,
since this is a complex double-edged effect and it should both feel worth
it AND dangerous).

(I also plan to replace Makhleb wrath with use of the Crucible so that we
can fully eliminate all the unsatisfying direct-damage wrath effects that
tend to either do nothing or kill the player with little counterplay).

Massive credit to regret-index for designing the portal vault (with heavy
subvaulting randomization), as well as the Crucible wall tiles and tiles
for slaughter form. Couldn't have done this without her.