Fix a crash upon using Vessel of Slaughter (Lightli)
This one is sort of amusing, since it didn't crash for me locally. But
apparently if you try to call zappy to fill out a beam struct for a zap
with no player defined hit/damage values, and are in wizmode, it will
*silently* replace those empty values with the monster equivalent zap. But
outside of wizmode, it will not (and then promptly assert due to not having
any values).

And of *course* I do my testing and development in wizmode... >.>