Have Beogh-pacified Orc drop their equipment upon leaving level (Acrobat)
Automatically pacifying orcs who were wearing equipment the player wanted
could be a sometimes-frustrating experience for Beogh worshippers. The
heavy penance Beogh exacts for attacking neutral orcs was intended to
discourage the deeply unflavorful act of murdering your converts to steal
their stuff, but in practice it was still sometimes possible to body-block
their exit from the level and find something else to kill them for you.

I'm not sure exactly how often this was practical, or how commonly players
actually did this, but at the same time I don't know that 'randomly lose
access to a monster's items from time to time' is an especially interesting
thing to try and preserve. So now pacified orcs will simply drop all their
items when they leave the level.

I've tried to flavor this in the most sensible way I could think of (since
one assumes they're not completely giving up the sword to live in the woods
and still have a use for those items themselves!). But you *are* the
possible messiah after all...