Some experimental projectile & cloud effect tile placeholders
These new tiles are mostly meant to try and reduce Crawl's use of the
exact same single cyan blast for effects themed after translocations,
metal, lightning, and necromantic effects, alongside a few others.
These new effects are for:

 * Metal Splinters (using roctavian's Sandblast)
 * Splinterspray and Volley of Thorns (also using that Sandblast)
 * Force Lance (using snw-0's spell icon for Force Lance)
 * Harpoon Shot (using roctavian's Stone Arrow and Iron Shot effects)
 * Ghostly Fireball (using unused Warp Weapon, Guardian Statue, and
                     Kitsune tiles by Denzi, jpeg, and amcnicky)
 * Calcifying Dust (using snw-0's seething chaos clouds)

(The Ghostly Fireball looks a little weird for a fireball but fine for a
negative energy effect, and that spell already probably should be joined
back to other negative energy themes if it's not going to have the healing
effect- even will-o-the-wisps didn't end up becoming undead. This still
leaves out Flashing Balestra and Momentum Strike in terms of keeping the
blast mostly for lightning effects, but I don't know what to use for the
former and the latter is part of a couple different effects that should be
adjusted to actually draw the beam at all times instead of only when it
misses.)