Refactor and simplify how monsters follow the player across stairs
The code that handled determining which monsters would follow with the
player when they took stairs or transporters had some opaque and
complicated behavior which may have once served an important purpose, but
enough of the game has changed around it that I think a simpler and more
intuitive method could be better.

Namely: friendly monsters could follow the player from up to 5 tiles away
from the staircase, but only so long as they were connected in a contiguous
chain, and *only* if there was not a single non-friendly monster adjacent
to the player at the same time (at which point the max distance allies
could be pulled from was 1). Note that 'non-friendly monster' in this case
included PLANTS. So merely being near a plant would cause you to leave
allies behind which you'd normally expect to follow you.

Note also that this chain didn't care about line of sight, so you could
pull along things around multiple corners, just so long as they were all
touching each other.

There was additional behavior of flagging monsters as MF_TAKING_STAIRS even
when we knew they *couldn't* take stairs, just so they could count as links
in that chain to allow access to stair-taking allies who might be behind
them, and needing to juggle separate return values for whether that flag
*really* meant they were taking the stairs or not. (Also, if one of them
broke the chain during transit, the rest of your allies wouldn't arrive...)

The original stated purpose of this code was to prevent zombies and summons
from blocking perma-allies behind them from travelling with the player, but
in the time since it was written, interlevel-recall was added and most
perma-allies removed from the game. There just isn't a compelling need for
this additional complexity anymore, in my opinion.

The new logic is: stair-using allies can follow from 3 tiles away (so long
as the player can see them) and adjacencies don't matter. MF_TAKING_STAIRS
should now only be used for monsters who are *actually* taking the stairs,
and determining followers can be done via a simple radius_iterator instead
of a sort of breadth-first search.

This does allow you to pull allies that are clearly not adjacent to the
you through transporters, which are instant, but it's not like the previous
logic ever pretended the monsters were walking multiple tiles into the gate
under their own power in a single turn anyway. I think it should be fine.

(I have also fixed a few other minor bugs at the same time, such as it
being possible to seperate Dowan and Duvessa if one of them gets confused
after you start taking the stairs but before you finish.)

Finally, since we no longer need to keep non-stair-using allies alive
during the entire transit process merely to determine which monsters WILL
transit, I have moved the code to kill such summons before it, which can
help with certain bookkeeping.