Prevent recasting Soul Splinter on a monster with a wisp already out
Or at the very least, say that they're not susceptible and prevent it from
working (being beam-targeted makes it a little hard to outright prevent it
from being cast in the direction of an ineligable monster). This prevents
the current wierdness of it making weakness wear off the target when you
recast it (as well as strange message ordering).

Additionally give better messages when the wisp doesn't form due to a lack
of adjacent space (and likewise don't auto-target an enemy in such a
position).