Re-enable Duel in Sprint
There were some past concerns about it affecting the balance of certain
sprints by letting the player duck out of situations that were never
intended to allow a breather, but the current version of Duel mostly
doesn't allow using it to stall. It seems like it should be fine to me.