Tweak some mutation lua tests
There are now no species in Crawl without any innate mutations at all, so
it is not possible for the lua tests to rely on any playable species not
having any.

In fact, it turns out that mutation.lua *already* silently relied on you
being a species with no innate mutations, even though that had stopped
being true by default since humans were changed, and it was only due to
'cleanup' at the end of a previous test in the default ordering that it
didn't assert. (ie: it would already assert if you ran the test on its own)

I've changed mut_species.lua to restore the player to the species they were
at the start of the test rather than to any specific species (this feels
closer to the idea of 'cleanup' to me and is what is currently done with
the player's position.)

As for mutation.lua, I changed the asserts to only ensure that NON-innate
mutations are cleaned up (this was only done during random mutation tests
anyway, and not ones that cared about mutation conflicts). If we really
want to ensure a blank slate of innates for the purposes of tests, I think
we would be better off either creating a method specifically for this
pupose or a vanilla species specifically to be used for tests only (and
thus not depending on vagueries of design changes over the years).