Improve Mark of the Tyrant's hostile summon behaviour (Lightli)
Getting a hostile demon from Infernal Servant also raises the quality of
the friendly demon you get, but with 25+ invo and Mark of the Tyrant, this
made demon qualify fall off the top of the chart and summon the default
demon (a red devil) instead.

Additionally, being able to get actual hostile fiends at the highest levels
of invo with this mark could feel overly punishing for the player's
investment, so let's cap hostiles to the same level that they would be at
high invo without having this mark.