Don't display Charms/Transmutations skills in Djinn dumps (inflori)
When manually enabling spell skill training on a djinn (either for a
character type that started with it off, or toggling it off and on again
on some other character), it would enable *all* magic skills, including
those which had been removed. I don't think had much effect, since it
wouldn't train them or display them in the skill menu, but they WOULD show
up in chardumps after that point.