Idiosyncratic earthworks (pianoman523)
New barreling boulder tiles, a more convincing Charlatan's Orb, and
something heavily repurposed into providing unique stone walls for the
Abyss. Abyss as a result has one of three different tile colours of stone
each game, which is a little weird but looks unique enough that I'm willing
to take the chance on this experiment.