Place duel opponents properly if you have LoS range < 3 (inflori)
Duel tries to place your opponent a random 3-5 tiles away from the player,
in a spot that the player still can see. But it is possible for kobold with
scarf of shadows and Robe of Night to have LoS 2, making those two
conditions mutually exclusive. This resulted in the monster placing
'anywhere' instead, which could include in the spectator stands where it
was impossible to reach them....

Now the min placement distance will be capped by you.current_vision