Remove a very tiny chance to fail to place a Duel opponent
monster::find_home_anywhere() (used as the default placement method for
transiting monsters) isn't guaranteed to actually work. It tries 600 times
to place a monster (choosing a random cell anywhere on the floor, which may
be well outside the bounds of a portal vault) and afterward just gives up.

This may be acceptable behavior for shafted monsters (though I am a little
dubious even here), but in the case of Duel it results in the Duel
immediately ending and the *next* time the player duels, that monster can
appear alongside the new one.

I'm not 100% sure this is the underlying cause of a recent bug, but it I
can't see any other obvious cause, and this definitely is *possible* (if
fantastically rare).

Anyway, just place the duel target next to the player on arrival (like with
stairs), which is more certain to work. (It will be relocated after that
time anyway.)

(Side note: monster::find_home_anywhere() doesn't appear to respect
no-tele-into either, meaning it is probably possible to shaft monsters into
display case vaults. Fixing this is somewhat non-trivial, though, since
I think the most obvious implementation would also prevent monsters from
taking stairs or hatches into no-tele-into areas as well, and I think at
least some vaults make use of this at present?)