Add "muscles" to the list of internal body parts.
So an octopode (with no bones) in death-form (with no blood) has one
singular internal body part (the soul) and one plural internal body
part (the muscles), so that internal body parts always have a possible
option.

External body parts are covered in singular by the head, and in plural
by the hands, so they always have a possible option already.