Tweak a hell message.
We don't know if the player can taste things, and while there is a
can_smell() function, there's no can_taste() function, and the sense of
taste is mostly the sense of smell anyway if real-life rules apply here.

As for the player's not having blood, the blood definitely isn't theirs
in that case, so the message still works for them.