Display future XL-based mutations on the 'A' screen (elliptic)
This commit will make the 'A' screen display mutations that will be gained
predictably at a future XL (ie: *not* for demonspawn or base draconians,
but for all other species-based mutations).

I have attempted to word these to make it somewhat more clear that the
player does not have the mutation in question *yet* (ie: turning 'you can'
into 'you will be able to' and so forth.) Simple automatic string
replacements cannot reliably shift the tense of everything, so a custom
message to describe a future mutation can also be specified in
mutation-data.h (and this is done for a few of them, including ones that
are especially long or would be really confusing, like claiming that
vampires will 'become afflicted by vampirism' by XL 3 because bat form is
technically an uprade to the vampirism mutation...

I'm not entirely happy with the implementation, but it should still be a
mild status quo improvement.