Clear the pursuers list after uncontrolled blinks/teleports
This mostly addresses the rare case of a teleport going off (or dispersal
trap being triggered) on a turn where the player had stepped away from some
monster and this translocation moving them into range of the monster when
they otherwise would not have been. (ie: it is possible to recieve an
attack of opportunity if you step away from a monster, teleport, and then
land 2 tiles away from their present location - even if you were nowhere
near the monster when you moved).

(Technically this makes the player immune to attacks of opportunity even in
cases where they were already in range and shifted only a single tile, but
that seems harmless.)