Make Blazeheart Golem explosion *actually* scale with spellpower.
This has been bugged since the spell was first implemented and would always
cause a 3d20 explosion regardless of spellpower (oh, the irony....). After
so long seeming to not be wildly overpowered at this level, I was a little
hesitant to nerf it to the originally-intended values, so I have moved
those upward a little. It will now reach 3d20 at 80 spellpower instead of
100 (capping at 3d22 at 100 power).

I expect it will still feel like it packs a sufficient punch (but it really
was intended to care about spellpower.)