Let monsters cast Grave Claw
Also mostly for ghosts and illusions. (Most of the code was already in
place for this to work earlier; I just hadn't done the last bits.)

Monsters, of course, are flagrant cheaters and don't have to care about
ammunition for this spell. Ghosts, at least, are probably as full of the
malice of the recently dead as they can hold.

(This means the spell description could be a little misleading with the
latter paragraph; in future I will try to find a way to fix that.)