Change how aux attack accuracy is displayed
Previously, the small description of aux attack mutations would list
trigger chance, accuracy, and base damage in one block, which misleadingly
made it look like the accuracy was in some way a property of that specific
aux attack. But in fact, all aux attacks have the same accuracy and this
was just reporting the player's current to-hit with aux attacks in
*general* (a fact that was misunderstood by literally everyone involved in
a recent discussion about them) and additionally doing so using a scale
of plusses that no one had any confidence about the meaning of.

We already report player melee to-hit in monster descriptions in a
human-readable way, so I am adding a description for aux attack accuracy
there as well (if the current character has any) and removing the line from
aux mutation descriptions.

I tried to be careful that the math here was correct (there are multiple
subtle differences between how the to-hit value is used by auxes versus
normal melee attacks) but am still not 100% confident.